Image functionality

+++ Eventopia.Infra/Service/CategoryService.cs
@@ -1,6 +1,8 @@
 ﻿using Eventopia.Core.Data;
 using Eventopia.Core.Repository;
 using Eventopia.Core.Service;
+using Eventopia.Infra.Utility;
+using Microsoft.AspNetCore.Http;
 
 namespace Eventopia.Infra.Service;
 
@@ -20,22 +22,42 @@ public class CategoryService : ICategoryService
 
 	public bool Delete(int id)
 	{
+		Category category = _categoryRepository.GetById(id);
+		if(category == null)
+			return false;
+		ImageUtility.DeleteImage(category.ImagePath, "Category");
 		return _categoryRepository.Delete(id);
 	}
 
 	public List<Category> GetAll()
 	{
-		return _categoryRepository.GetAll();
+		List<Category> categories = _categoryRepository.GetAll();
+		foreach(var category in categories)
+		{
+			byte[]? byteFile = ImageUtility.RetrieveImage(category.ImagePath, "Category");
+			category.RetrievedImageFile = byteFile;
+        }
+		return categories;
 	}
 
 	public Category GetById(int id)
 	{
-		return _categoryRepository.GetById(id);
+		Category category = _categoryRepository.GetById(id);
+		if (category == null)
+			return null;
+		byte[]? byteFile = ImageUtility.RetrieveImage(category.ImagePath, "Category");
+		category.RetrievedImageFile = byteFile;
+		return category;
 	}
 
 	public Category GetCategoryByName(string name)
 	{
-		return _categoryRepository.GetCategoryByName(name);
+		Category category = _categoryRepository.GetCategoryByName(name);
+		if (category == null)
+			return null;
+		byte[]? byteFile = ImageUtility.RetrieveImage(category.ImagePath, "Category");
+		category.RetrievedImageFile = byteFile;
+		return category;
 	}
 
 	public bool Update(Category category)
